requirements.txt
chromadb>=0.5.0
sentence-transformers>=3.0.0
faiss-cpu>=1.7.4
numpy>=1.26
pymilvus>=2.4.4
